Restore logging code
